Add statuses

+++ Assets/Scripts/Main/UI/Presenters/StartWindowPresenter.cs
@@ -133,17 +133,46 @@ namespace Main.UI.Presenters {
             if (!_globalScope.ApprovedMatchAndNeedLoad) return;
             _globalScope.ApprovedMatchAndNeedLoad = false;
 
+            if (!_globalScope.SendedInvites.ContainsKey(_appConfig.OpponentUserId)) return;
+            
             var inviteData = _globalScope.SendedInvites[_appConfig.OpponentUserId];
             _appConfig.OpponentDisplayName = inviteData.DisplayName;
-            foreach (var pair in _globalScope.SendedInvites) {
-                
-            }
+            await DeclineAllReceivedSignals();
+            await DeclineAllSendedSignals();
             _globalScope.SendedInvites.Clear();
             await _nakamaService.RemoveAllPartiesExcept(_appConfig.OpponentUserId);
             await LoadParty();
         }
+        
+        private async UniTask DeclineAllSendedSignals() {
+            UniTask[] tasks = new UniTask[_globalScope.SendedInvites.Count];
+            int i = 0;
+            foreach (var pair in _globalScope.SendedInvites)
+            {
+                tasks[i] = _messageService.SendDeclineInviteSended(pair.Key);
+                i++;
+            }
 
-        private void CheckInvite() {
+            _globalScope.SendedInvites.Clear();
+            await UniTask.WhenAll(tasks);
+        }
+
+        private async UniTask DeclineAllReceivedSignals()
+        {
+            UniTask[] tasks = new UniTask[_globalScope.ReceivedInvites.Count];
+            int i = 0;
+            foreach (var pair in _globalScope.ReceivedInvites)
+            {
+                tasks[i] = _messageService.SendDeclineInviteReceived(pair.Key);
+                i++;
+            }
+            
+            _globalScope.ReceivedInvites.Clear();
+
+            await UniTask.WhenAll(tasks);
+        }
+
+        private async void CheckInvite() {
             if (_globalScope.ReceivedInvites.Count == 0) return;
             if (_windowService.IsWindowOpened(WindowKey.InviteWindow)) return;
 
@@ -156,6 +185,9 @@ namespace Main.UI.Presenters {
 
             _globalScope.ReceivedInvites.Remove(inviteData.Key);
 
+            var userData = await _nakamaService.GetUserInfo(inviteData.Key);
+            if (!userData.Online) return;
+            
             _signalBus.Fire(new OpenWindowSignal(WindowKey.InviteWindow, new InviteWindowData {
                 InviteData = inviteData.Value
             }));
